Updates BLE services and adds battery measurement functionality
This commit updates the BLE services by adding support for the ANT Scan Service and the Battery Service. It includes changes to the `.vscode/settings.json` file and the `pca10056/s340/armgcc/Makefile`. The commit also introduces the `ble_battery_service.c` and `ble_battery_service.h` files for implementing the Battery Service. The `ble_ant_scan_service.c` and `ble_ant_scan_service.h` files are modified to remove unused constants. Additionally, the `ble_setup.c` file is updated to include the new services and the `battery_measurement.c` and `battery_measurement.h` files are added for battery voltage measurement. This commit improves the functionality of the BLE services and allows for battery level monitoring in the application.