Adds support for sending current ANT+ Device ID as a notification
Introduces a new function `send_current_ant_device_id` to send the current ANT+ Device ID as a notification. This function is called when a specific write command (0x04) is received. The current ANT+ Device ID is split into two bytes and sent as data in the notification. The commit also includes relevant logging statements to indicate the success or failure of the notification.