Adds BLE delay timer for device shutdown
Introduces a new BLE delay timer to handle device shutdown when there is no BLE connection after 20 seconds. If BLE is still not connected, the system shuts down by closing the ANT+ channel, enabling the reed switch, and entering deep sleep. If BLE is connected, the shutdown is delayed for an additional 20 seconds. This change improves the shutdown process and ensures proper handling of BLE and ANT+ channels.