Adds build_all target and bootloader build support
Introduces a new build_all target in the Makefile to build both the bootloader and the default firmware. Also adds support for building the bootloader in the build_bootloader target.

This commit improves the build system by providing a convenient way to build both the bootloader and the firmware in a single step.