Adds support for BLE battery service and updates FTMS and CPS services
This commit adds support for the BLE battery service by including the "ble_battery_service" files, which consist of the "ble_battery_service.c" and "ble_battery_service.h" files. Additionally, the FTMS and CPS services have been updated to use the new battery service. This change allows for better management and monitoring of the battery level and power state.